refactor(e2e): extract shared expectValidExportZip helper
Move duplicate export zip validation logic from chart and dataset
spec files into a shared helper in assertions.ts. The helper takes
a resourceDir parameter ('charts' or 'datasets') to validate the
correct directory structure within the zip.
